requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.1.12
gunicorn==21.2.0
packaging==23.2
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
whitenoise==6.6.0